requirements.txt
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
diskcache==5.6.3
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
llama_cpp_python==0.3.7
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.20.1
packaging==24.2
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
av==14.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
ctranslate2==4.5.0
diskcache==5.6.3
faster-whisper==1.1.1
filelock==3.17.0
flatbuffers==25.2.10
fsspec==2025.2.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
keyboard==0.13.5
llama_cpp_python==0.3.7
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.3
onnxruntime==1.20.1
OPi.GPIO==0.5.2
packaging==24.2
protobuf==5.29.3
PyYAML==6.0.2
requests==2.32.3
sympy==1.13.3
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
